+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed HLS tracks selection.
 * Removed HLS parser package and included HLS parser package in Better Player.
 * Removed unused player observer in iOS.
+* Fixed cache issue in Android where multiple Better Player instances uses same directory.
 
 ## 0.0.52
 * Fixed unregister listener issue in iOS.